// fc-monolito/src/modules/invoice/usecase/find-invoice/find-invoice.dto.ts
export interface FindInvoiceInputDto {
  id: string;
}

export interface FindInvoiceOutputDto {
  id: string;
  name: string
  document: string
  street: string;
  number: string;
  complement: string;
  city: string;
  state: string;
  zipCode: string;
  items: {
    id: string;
    name: string;
    price: number;
  }[];
  total: number;
  createdAt?: Date;
  updatedAt?: Date;
}